PhanMemKeToan_backend/requirements.txt
# Core FastAPI framework
fastapi==0.104.1

# ASGI server
uvicorn[standard]==0.24.0

# Database ORM
sqlalchemy==2.0.23

# PostgreSQL adapter
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.5.0

# JWT authentication
PyJWT==2.8.0

# Password hashing
Werkzeug==3.0.1

# CORS middleware
python-multipart==0.0.6

# Date/time utilities
python-dateutil==2.8.2

# Logging
colorlog==6.7.0 
